+++ src/main/java/be/ugent/rml/cli/Main.java
@@ -27,6 +27,15 @@ public class Main {
     private static final Logger logger = LoggerFactory.getLogger(Main.class);
 
     public static void main(String[] args) {
+        main(args, System.getProperty("user.dir"));
+    }
+
+    /**
+     * Main method use for the CLI. Allows to also set the current working directory via the argument basePath.
+     * @param args the CLI arguments
+     * @param basePath the basePath used during the execution.
+     */
+    public static void main(String[] args, String basePath) {
         Options options = new Options();
         Option mappingdocOption = Option.builder("m")
                 .longOpt("mappingfile")
@@ -125,7 +134,7 @@ public class Main {
                         .collect(Collectors.toList());
                 InputStream is = new SequenceInputStream(Collections.enumeration(lis));
                 RDF4JStore rmlStore = Utils.readTurtle(is, RDFFormat.TURTLE);
-                RecordsFactory factory = new RecordsFactory(new DataFetcher(System.getProperty("user.dir"), rmlStore));
+                RecordsFactory factory = new RecordsFactory(basePath);
 
                 String outputFormat = getPriorityOptionValue(serializationFormatOption, lineArgs, configFile);
                 QuadStore outputStore;